put the spatialite loader back to mjolnir/util.h and remove spatialite_conn.h/cc
