+++ src/mjolnir/CMakeLists.txt
@@ -43,8 +43,7 @@ set(sources
   servicedays.cc
   speed_assigner.h
   timeparsing.cc
-  util.cc
-  spatialite_conn.cc)
+  util.cc)
 
 set (sources_with_warnings
   admin.cc
