+++ valhalla/mjolnir/util.h
@@ -8,6 +8,10 @@
 #include <unordered_map>
 #include <vector>
 
+#include <sqlite3.h>
+// needs to be after sqlite include
+#include <spatialite.h>
+
 #include <valhalla/baldr/graphid.h>
 #include <valhalla/baldr/rapidjson_utils.h>
 #include <valhalla/midgard/logging.h>
@@ -122,6 +126,14 @@ bool shapes_match(const std::vector<midgard::PointLL>& shape1,
  *         more curved and tighter turns.
  */
 uint32_t compute_curvature(const std::list<midgard::PointLL>& shape);
+
+/**
+ * Will allocate a spatialite connection
+ *
+ * @param handle The sqlite database handle
+ * @return a shared pointer to the connection which cleans up after itself when destructed
+ */
+std::shared_ptr<void> make_spatialite_cache(sqlite3* handle);
 
 /**
  * Build an entire valhalla tileset give a config file and some input pbfs. The
